+++ ldap2pg/ldap.py
@@ -46,13 +46,44 @@ def parse_scope(raw):
         raise ValueError("Unknown scope %r" % (raw,))
 
 
-if PY2:  # pragma: nocover_py3
-    def str2dn(value):
-        # Workaround buggy unicode managmenent in upstream python-ldap. This is
-        # not necessary with pyldap on Python3.
-        return decode_value(native_str2dn(value.encode('utf-8')))
-else:  # pragma: nocover_py2
-    str2dn = native_str2dn
+def str2dn(value):
+    try:
+        if PY2:  # pragma: nocover_py3
+            # Workaround buggy unicode managmenent in upstream python-ldap.
+            # This is not necessary with pyldap on Python3.
+            return decode_value(native_str2dn(value.encode('utf-8')))
+        else:  # pragma: nocover_py2
+            return native_str2dn(value)
+    except ldap.DECODING_ERROR:
+        raise ValueError("Can't parse DN '%s'" % (value,))
+
+
+def get_attribute(entry, attribute):
+    _, attributes = entry
+    path = attribute.split('.')
+    try:
+        values = attributes[path[0]]
+    except KeyError:
+        raise ValueError("Unknown attribute %r" % (path[0],))
+    path = path[1:]
+    for value in values:
+        if path:
+            try:
+                dn = str2dn(value)
+            except ValueError:
+                msg = "Can't parse DN from attribute %s=%s" % (
+                    attribute, value)
+                raise ValueError(msg)
+            value = dict()
+            for (type_, name, _), in dn:
+                names = value.setdefault(type_, [])
+                names.append(name)
+            try:
+                value = value[path[0]][0]
+            except KeyError:
+                raise ValueError("Unknown attribute %s" % (path[0],))
+
+        yield value
 
 
 class EncodedParamsCallable(object):  # pragma: nocover_py3