+++ tests/unit/test_ldap.py
@@ -13,6 +13,16 @@ base  ou=IT staff,o="Example, Inc",c=US
 """
 
 
+def test_str2dn():
+    from ldap2pg.ldap import str2dn
+
+    value = str2dn('cn=toto,dc=with space,dc=pouet')
+    assert 3 == len(value)
+
+    with pytest.raises(ValueError):
+        str2dn('not a dn')
+
+
 def test_connect_from_env(mocker):
     go = mocker.patch('ldap2pg.ldap.gather_options', autospec=True)
     li = mocker.patch('ldap2pg.ldap.ldap.initialize', autospec=True)
@@ -123,15 +133,21 @@ def test_parse_rc():
     assert 3 == items[0].lineno
 
 
-def test_get_ldap_attr():
-    from ldap2pg.manager import get_ldap_attribute
+def test_get_attribute():
+    from ldap2pg.manager import get_attribute
 
     with pytest.raises(ValueError):
-        list(get_ldap_attribute(entry=('dn', {}), attribute='pouet'))
+        list(get_attribute(entry=('dn', {}), attribute='pouet'))
 
     with pytest.raises(ValueError):
-        list(get_ldap_attribute(
+        list(get_attribute(
             entry=('dn', {'cn': ['cn=pouet']}),
+            attribute='cn.pouet',
+        ))
+
+    with pytest.raises(ValueError):
+        list(get_attribute(
+            entry=('dn', {'cn': 'not a dn'}),
             attribute='cn.pouet',
         ))